LayoutTests/svg/dynamic-updates/SVGFEBlendElement-svgdom-in2-prop-expected.txt

CONSOLE MESSAGE: line 215: TypeError: 'undefined' is not an object (evaluating 'normalBlend.in2.baseVal = "normalImg"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' property of the SVGFEBlendElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL normalBlend.in2.baseVal should be lightenImg. Threw exception TypeError: 'undefined' is not an object (evaluating 'normalBlend.in2.baseVal')
FAIL multiplyBlend.in2.baseVal should be normalImg. Threw exception TypeError: 'undefined' is not an object (evaluating 'multiplyBlend.in2.baseVal')
FAIL screenBlend.in2.baseVal should be multiplyImg. Threw exception TypeError: 'undefined' is not an object (evaluating 'screenBlend.in2.baseVal')
FAIL darkenBlend.in2.baseVal should be screenImg. Threw exception TypeError: 'undefined' is not an object (evaluating 'darkenBlend.in2.baseVal')
FAIL lightenBlend.in2.baseVal should be darkenImg. Threw exception TypeError: 'undefined' is not an object (evaluating 'lightenBlend.in2.baseVal')
